requirements.txt
configparser
mendeley
ordereddict
pdfminer3k
wheel
